axi_spi_engine: Delete remove_bus_interface calls
This version of Vivado infers better the interfaces. No need to remove
these three interface, because it will not be inferred automatically.
If we do not delete the ipx command calls, none of the SPI interface will
be inferred for some reason. And that will result breaking every project,
that uses the SPI Engine.