Makefile: Don't create invalid sub-project targets
At the moment targets are created for any file that is in a sub-directory
of the project directory. This is not the intended behavior. Targets should
only be generated for sub-projects. We can identify those by filtering for
sub-directories that have a Makefile.

E.g. some of those invalid targets were "adv7511.Makefile" and
"scripts.adi_board.tcl".

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>