jesd204_soft_pcs_tx: Add support for lane polarity inversion
Some designs choose to swap the positive and negative side of the of the
JESD204 lanes. One reason for this would be because it can simplify the
PCB layout.

To support this add a parameter to the jesd204_soft_pcs_tx module that
allows to specify whether the lane polarity is inverted or not.

The way the polarity inversion is implemented is for free since it just
inverts the output mapping of the 8b10b encoder LUT tables.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>