axi_dmac: diagnostic interface in beats
This change adds to the diagnostic interface
fullness of the internal buffer represented in beats.
The signal is driven from the destination clock domain,
as this is reflected in its name.

The signal has a fixed size and is dimensioned by
taking in account the supported maximum number of bursts of 128 and
burst length of 256 beats.

Since the DMAC is working in burst the counter has
an uncertainity of 0..beats_per_burst
meaining that at a certain point the real number of beats that are
available for destination can be smaller than the reported number.
If the source side of the DMAC stalls, the incomplete burst won't be
sent to the destination.