ad_ip_jesd204_tpl_adc: Remove unnecessary ad_xcvr_rx_if
The ad_xcvr_rx_if module aligns the incoming data stream so that the first
octet in a frame is in the LSBs of the data word.

But the data that comes from the JESD204 link layer is already aligned so
that the first octet of a frame is in the LSBs. So this is unnecessary, it
just uses up resources and adds additional latency. Remove it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>