Add generated files for Intel projects to .gitignore
The system_db folder is autogenerated and contains all the files that are
generated by the Platform Designer tool.

The extension for Intel binary bitstreams is rbf.

Add both of those to the .gitignore since they should not be under version
control and just end up as clutter in `git status` otherwise.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>